feat: add popular contest chart data to contest statistics API

+++ core-service/core-domain/core-application-service/src/main/java/com/backend/programming/learning/system/core/service/domain/dto/responseentity/contest/QueryLineChartResponse.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 @Builder
 public class QueryLineChartResponse {
     @JsonProperty("data")
-    private final float[] data;
+    private final double[] data;
 
     @JsonProperty("label")
     private final String label;
